FIX: move test API to separate file
