+++ backend/main.py
@@ -11,6 +11,8 @@ import hashlib
 import uuid
 import json
 import time
+from tests.test import router as test_webhook_router
+
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -41,9 +43,13 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-@app.get("/")    
-def say_hello():
-    return {"message": "Hello, world!"}
+
+# Include the test_webhook router
+app.include_router(test_webhook_router)
+
+# Set the supabase client in the test_webhook module
+import tests.test
+tests.test.supabase = supabase
 
 @app.post("/sync-user")
 async def sync_user(request: Request):
@@ -100,67 +106,6 @@ async def sync_user(request: Request):
         logger.error(f"Error in sync_user: {str(e)}")
         raise HTTPException(status_code=500, detail=str(e))
 
-
-@app.post("/test-webhook")
-async def test_webhook(request: Request):
-    """Test endpoint to simulate Clerk webhook for user deletion."""
-    try:
-        data = await request.json()
-        user_id = data.get("user_id")
-        
-        if not user_id:
-            return {"status": "error", "message": "Missing user_id parameter"}
-            
-        # Create a simulated webhook payload that exactly matches Clerk's format
-        webhook_payload = {
-            "data": {
-                "deleted": True,
-                "id": user_id,
-                "object": "user"
-            },
-            "event_attributes": {
-                "http_request": {
-                    "client_ip": "127.0.0.1",
-                    "user_agent": "Test Agent"
-                }
-            },
-            "object": "event",
-            "timestamp": int(time.time() * 1000),
-            "type": "user.deleted"
-        }
-        
-        # Log the payload
-        logger.info(f"Simulating Clerk webhook with payload: {webhook_payload}")
-        
-        # Check if user exists
-        check_user = supabase.table("users").select("*").eq("user_id", user_id).execute()
-        user_exists = check_user.data and len(check_user.data) > 0
-        
-        if not user_exists:
-            return {"status": "warning", "message": f"User {user_id} not found in Supabase, nothing to delete"}
-            
-        # Delete the user
-        response = (
-            supabase.table("users")
-            .delete()
-            .eq("user_id", user_id)
-            .execute()
-        )
-        
-        if hasattr(response, 'error') and response.error:
-            logger.error(f"Failed to delete user: {response.error}")
-            return {"status": "error", "message": str(response.error)}
-            
-        return {
-            "status": "success", 
-            "message": f"User {user_id} deleted successfully", 
-            "data": response.data,
-            "webhook_payload": webhook_payload
-        }
-        
-    except Exception as e:
-        logger.error(f"Error in test webhook: {str(e)}")
-        return {"status": "error", "message": str(e)}
 
 @app.post("/webhook/clerk")
 async def clerk_webhook(request: Request):
